Git custom commit message template
Prior to this change, a template for git commits did not exist.
Having a template such as this helps guide contributors towards
writing complete and cogent commit messages.

This commit adds a template for a git commit log. It adds this template
to the global git config file of the user.

On a miscellaneous note, the default tabstop for less was set to 4.