Remove existing git commit template
Prior to this change, if the init script had been run, an error would
have been thrown when the symlink was setup with the git commit
template.

This commit makes sure any existing git commit template of the name used
in this project is removed before the symlink is made.